apiとなる関数をapi.tsに分割しました。 pull/4#discussion_r1799399321への対応です。
