+++ src/App.tsx
@@ -1,14 +1,12 @@
 import './App.css'
-import { useState } from 'react'
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
+import { useCookies } from "react-cookie";
 import { GoogleMap, LoadScript } from "@react-google-maps/api";
 import { BrowserRouter, Link, Route, Routes,useNavigate, useLocation} from "react-router-dom";
 import Signup from './signup';
-import { useCookies } from "react-cookie";
+import api from './api/api';
 
 function App() {
-  const [userName, setUserName] = useState<string>("");
-  const [userEmail, setUserEmail] = useState<string>("");
   const [cookies,, removeCookie] = useCookies(["isSession"]);
   const isAuthenticated = !!cookies.isSession; // 認証されているかどうか
   const navigate = useNavigate(); // 画面遷移をするためにuseNavigate フックを使用
@@ -36,46 +34,9 @@ function App() {
     lng: 139.7528246814339,
   };
 
-  async function callGetStores() {
-    try {
-      const response = await fetch('http://localhost:8080/');
-      const data = await response.json();
-      console.log(data)
-    } catch (error) {
-      console.error('error:', error);
-    }
-  };
-
-  async function callCreateUser() {
-    try {
-      const userData = {
-        name: userName,
-        email: userEmail,
-      };
-      const response = await fetch('http://localhost:8080/user', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(userData),
-      });
-      const data = await response.json();
-      console.log(data)    
-    } catch (error) {
-      console.error('error:', error);
-    }
-  }
-
-  function handleChangeUserName(e: React.ChangeEvent<HTMLInputElement>) {
-    setUserName(e.target.value);
-  }
-
-  function handleChangeUserEmail(e: React.ChangeEvent<HTMLInputElement>) {
-    setUserEmail(e.target.value);
-  }
-
   return (
     <>
+    {/* ヘッダー部分 */}
     <div className="App">
       <Link to="/">Home</Link>
       <br />
@@ -87,21 +48,14 @@ function App() {
         <Route path="/signup" element={<Signup />} />
       </Routes>
     </div>
+
+
       <h1>Clean Storemap Web</h1>
       {(isAuthenticated)?"ログインしています。":"ログインしていません。"}
       <div className="card">
-        <button onClick={() => callGetStores()}>
+        <button onClick={() => api.callGetStores()}>
           押すぅ！
         </button>
-      </div>
-      <div>
-        name
-        <input value={userName} onChange={handleChangeUserName}/>
-        <p>{userName}</p>
-        email
-        <input value={userEmail} onChange={handleChangeUserEmail}/>
-        <p>{userEmail}</p>
-        <button onClick={() => callCreateUser()}>Create User</button>
       </div>
 
       <LoadScript googleMapsApiKey={import.meta.env.VITE_GOOGLE_MAPS_API_KEY}>
