+++ src/types/formData.ts
@@ -1,0 +1,10 @@
+// createUserでPost通信する時のdataの型
+interface FormData {
+  name: string;
+  email: string;
+  age: number;
+  sex: number;
+  gender: number;
+}
+
+export default FormData;
