--- src/utils/Request.tsx
@@ -1,6 +1,0 @@
-const requests = {
-  getStores: 'http://127.0.0.1:8080/',  // GETメソッド 
-  createUser: 'http://127.0.0.1:8080/user',  // POSTメソッド  
-};
-
-export default requests;
